Changed budget, revenue and id type from int to long

+++ app/src/main/java/com/dmikhov/cinemin/screens/MainActivity.kt
@@ -22,7 +22,7 @@ class MainActivity : AppCompatActivity() {
 
     }
 
-    fun openHomeFragment(movieId: Int?, movieTitle: String?) {
+    fun openHomeFragment(movieId: Long?, movieTitle: String?) {
         val homeFragment = MovieFragment.newInstance(movieId, movieTitle)
         supportFragmentManager.beginTransaction()
             .add(R.id.rootFragmentContainer, homeFragment)
