+++ entities/src/main/java/com/dmikhov/entities/movie/Movie.kt
@@ -1,13 +1,13 @@
 package com.dmikhov.entities.movie
 
 data class Movie (
-    val id: Int,
+    val id: Long,
     val title: String?,
     val posterUrl: String?,
     val releaseDate: String?,
-    val budget: Int?,
-    val revenue: Int?,
-    val runtime: Int?,
+    val budget: Long?,
+    val revenue: Long?,
+    val runtime: Long?,
     val rating: Float?,
     val adult: Boolean,
     val tagline: String?,
